Updated script
The review_time column was Unix timestamps that show the each year the given review was given. After the converting process it is datetime format and created a new column as review_year that include only year values.
Also, Added new process which is feature engineering. Thanks to this process, will go deeply to each column. It will provide better insight in following stages.